requirements.txt
#Installation Requirements
python-binance
TA-Lib
numpy
websocket_client